% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callr_wrappers.R
\name{bgjm_poll}
\alias{bgjm_poll}
\title{Poll background jobs}
\usage{
bgjm_poll(timeout_ms = 0)
}
\arguments{
\item{timeout_ms}{Milliseconds to wait inside \code{\link[callr:reexports]{callr::poll()}}.}
}
\value{
Invisible poll result from \code{\link[callr:reexports]{callr::poll()}}.
}
\description{
Poll background jobs
}
