% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callr_wrappers.R
\name{bgjm_read_stdout}
\alias{bgjm_read_stdout}
\title{Read the captured standard output}
\usage{
bgjm_read_stdout(job_id, n = -1L)
}
\arguments{
\item{job_id}{Identifier returned by a \verb{bgjm_start_*()} function.}

\item{n}{Number of lines to read; defaults to all lines.}
}
\value{
Character vector of log lines.
}
\description{
Read the captured standard output
}
