% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callr_wrappers.R
\name{bgjm_remove}
\alias{bgjm_remove}
\title{Remove a job from the registry}
\usage{
bgjm_remove(job_id, delete_dir = FALSE)
}
\arguments{
\item{job_id}{Identifier returned by a \verb{bgjm_start_*()} function.}

\item{delete_dir}{When \code{TRUE}, delete the job's working directory.}
}
\value{
Invisible logical indicating whether the job existed.
}
\description{
Remove a job from the registry
}
