% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callr_wrappers.R
\name{bgjm_status}
\alias{bgjm_status}
\title{Query job status}
\usage{
bgjm_status(job_id)
}
\arguments{
\item{job_id}{Identifier returned by a \verb{bgjm_start_*()} function.}
}
\value{
\code{"running"}, \code{"finished"}, or \code{"error"}.
}
\description{
Query job status
}
