% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franzpak-package.R
\docType{package}
\name{franzpak-package}
\alias{franzpak}
\alias{franzpak-package}
\title{franzpak: Miscellaneous functions for social science data analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Francisco Wilhelm \email{francisco.wilhelm@unibe.ch} (\href{https://orcid.org/0000-0002-1717-7975}{ORCID})

}
\keyword{internal}
