% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callr_wrappers.R
\name{\%||\%}
\alias{\%||\%}
\title{Return the left operand if not \code{NULL}}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side.}

\item{y}{Right-hand side.}
}
\value{
\code{x} when it is not \code{NULL}, otherwise \code{y}.
}
\description{
This is a lightweight infix helper used to provide default values.
}
\keyword{internal}
