% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_items}
\alias{mc_items}
\title{Monte Carlo-simulated dataset with continuous factor indicators}
\format{
\subsection{\code{mc_items}}{

A data frame with 500 rows and 10 columns:
\describe{
\item{firstscale_1, firstscale_2, firstscale_3, firstscale_4, firstscale_5}{indicators of first factor}
\item{secondscale_1, secondscale_2, secondscale_3, secondscale_4, secondscale_5}{indicators of second factor, with secondscale_5 being negatively related to factor.}
}
}
}
\source{
\url{https://www.statmodel.com/usersguide/chapter12.shtml}
}
\usage{
mc_items
}
\description{
A dataset generated through monte carlo simulation. See Mplus 8 example 12.1.
}
\keyword{datasets}
