% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_twolevel}
\alias{mc_twolevel}
\title{Monte Carlo-simulated two-level dataset}
\format{
\subsection{\code{mc_twolevel}}{

A data frame with 1,000 rows and 4 columns:
\describe{
\item{Y}{dependent variable}
\item{X, M}{independent and mediating variable}
\item{CLUSTER}{gives the cluster ID}
}
}
}
\source{
\url{https://www.statmodel.com/usersguide/chapter12.shtml}
}
\usage{
mc_twolevel
}
\description{
A dataset generated through monte carlo simulation. See Mplus 8 example 12.11.
}
\keyword{datasets}
