% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_longer_multilevel.R
\name{pivot_longer_multilevel}
\alias{pivot_longer_multilevel}
\title{Gather wide two-level data into long format
Specifically designed for two-level data of repeated measures. Will create NAs for missing time points.}
\usage{
pivot_longer_multilevel(data, varlist, maxlevel1, varnames = NULL)
}
\arguments{
\item{data}{dataframe containing two-level data in wide format}

\item{varlist}{list of variables to be pivoted into longer format}

\item{maxlevel1}{specify the maximum number of level 1 measurements per cluster.}

\item{varnames}{optionally specify the names of the new variables, otherwise names are taken from varlist.}
}
\value{
a tibble with id (identifying the level-2 unit), time (identify the ordered level-1 unit), and further columns as specified in varlist (or varnames, if given).
}
\description{
Gather wide two-level data into long format
Specifically designed for two-level data of repeated measures. Will create NAs for missing time points.
}
\details{
The function assumes that the sequence of level 1 measurements is ordered and meaningful, as is the case with repeated measurements / diary designs.
This contrasts with multilevel data gathered from groups (e.g., organizations) in which the sequence of level 1 meaurements is unordered and incidental.
However, the function may still be useful for data with unordered level-1 measurements.
The maximum number of rows of the returned tibble is id (number of level-2 observations) x maxlevlel1 (maximum number of level-1 observations. Completely NA level-1 observations are removed, which may result in a reduced number of rows.
}
\examples{
# generate wide dataset with 10 clusters, x measured three times and y measured three times
df_twolevel_wide <- tibble::tibble(id = 1:10, x_t1 = rep(1:5,2),
  x_t2 = rep(1:5,2), x_t3 = rep(1:5,2),
  y_t1 = rep(1:5,2), y_t2 = rep(1:5,2), y_t3 = rep(1:5,2))
pivot_longer_multilevel(df_twolevel_wide, c("x", "y"), 3)
}
