% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreItemsMulti.R
\name{scoreItemsMulti}
\alias{scoreItemsMulti}
\title{Automatically score multiple scales and return several psychometrical scale parameters}
\usage{
scoreItemsMulti(scalenames, dataframe, exclude = TRUE, manual_keys = NULL)
}
\arguments{
\item{scalenames}{character vector of scale names}

\item{dataframe}{dataframe holding the items to be scored}

\item{exclude}{Boolean indicating whether to exclude participant responses where more than 1/3 of a scale are NA.}

\item{manual_keys}{named list with manual keys, formatted like in scoreItems.}
}
\value{
a list object holding scale scores and other information
}
\description{
This is a convenience wrapper for psych::scoreItems that allows automatically
scoring multiple scales at once.
}
