% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_coef_tools.R
\name{sep_label}
\alias{sep_label}
\title{Separate Mplus output labels into X and Y (internal function)}
\usage{
sep_label(model)
}
\arguments{
\item{model}{mplus model as input.}
}
\value{
dataframe with coefficients and DV (dependent variable) and IV (independent variable) labels.
}
\description{
Separate Mplus output labels into X and Y (internal function)
}
