% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{star_assign}
\alias{star_assign}
\title{Assign significance signs following APA nomenclature}
\usage{
star_assign(x)
}
\arguments{
\item{x}{}
}
\description{
Assign significance signs following APA nomenclature
}
