% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callr_wrappers.R
\name{bgjm_collect}
\alias{bgjm_collect}
\title{Collect the result of a background job}
\usage{
bgjm_collect(job_id, read_rds = TRUE, auto_remove = FALSE)
}
\arguments{
\item{job_id}{Identifier returned by a \verb{bgjm_start_*()} function.}

\item{read_rds}{When \code{TRUE}, read serialized results produced with the \code{"rds"}
strategy.}

\item{auto_remove}{If \code{TRUE}, remove the job from the registry after
collection.}
}
\value{
A list containing the job metadata, logs, and result object.
}
\description{
Collect the result of a background job
}
