% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callr_wrappers.R
\name{bgjm_kill}
\alias{bgjm_kill}
\title{Terminate a running job}
\usage{
bgjm_kill(job_id)
}
\arguments{
\item{job_id}{Identifier returned by a \verb{bgjm_start_*()} function.}
}
\value{
Invisible \code{TRUE} when the signal is sent successfully.
}
\description{
Terminate a running job
}
