% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_coef_tools.R
\name{coef_wrapper}
\alias{coef_wrapper}
\title{Format Mplus Model Coefficients}
\usage{
coef_wrapper(
  model,
  label_replace = NULL,
  params = c("regression"),
  bayes = FALSE,
  addci = FALSE
)
}
\arguments{
\item{model}{the Mplus model from MplusAutomation.}

\item{label_replace}{named character vector for replacing parameter labels}

\item{params}{which types of parameters to extract, see MplusAutomation}

\item{bayes}{Double p-value for one-tailed Bayes tests (gives rough p-estimates)}

\item{addci}{Add confidence intervals to parameters}
}
\value{
a dataframe with mplus model coefficients
}
\description{
Format Mplus Model Coefficients
}
