% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortable_multilevel.R
\name{cortable_multilevel}
\alias{cortable_multilevel}
\title{Correlation table for two-level data}
\usage{
cortable_multilevel(df, varnames, grp)
}
\arguments{
\item{df}{the dataframe used}

\item{varnames}{the variable names as string vector}

\item{grp}{the variable name that identifies the level-2 group as string character}
}
\value{
a correlation table as tibble
}
\description{
Correlation table for two-level data
}
